pyproject.toml
[tool.poetry]
name = "headlamphealth"
version = "0.1.0"
description = ""
authors = ["abhinavkashyap <abhinav@comp.nus.edu.sg>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.13"
pandas = "^2.2.2"
rich = "^13.7.1"
scikit-learn = "^1.5.0"
streamlit = "^1.35.0"
langchain-community = "^0.2.4"
langchain-chroma = "^0.1.1"
streamlit-card = "^1.0.2"
langchain-openai = "^0.1.8"
python-dotenv = "^1.0.1"
sentence-transformers = "^3.0.1"
langchain-huggingface = "^0.0.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
